Hands_on_Lab/Day_2/Task_4/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
pypdf
groq
torch
torchvision
torchaudio
huggingface-hub
tiktoken
